kernel: synchronize with the latest 'scope-minimized manual hooks'
Description addendum:
kernel: ksud, core_hook: migrate ksud execution to security_bprm_check
https://github.com/pershoot/KernelSU-Next/commit/c2e2e4773830b89df188a965bf2c73035f8fced4

as for hardening:
...
- copy_from_user_nofault + a copy_from_user fallback

My reasoning on adding a fallback is that on some devices a fault happens,
and it is copied garbled data. On my creation of this, I actually had to
lock that _nofault copy on a spinlock as a way to mimic preempt_disable/enable
without actually doing it. As per user reports, no failed _nofault copies anyway
but we have-to-have a fallback for resilience.

References:
- old version https://github.com/backslashxx/KernelSU/commit/6efcd8193e62d13a4e62cda0ce54d6770260c680
- bad usercopy https://github.com/backslashxx/KernelSU/issues/21

This now provides a small helper function, ksu_copy_from_user_retry, which explains
itself. First we attempt a _nofault copy, if that fails, we try plain.

With that, I also need to provide  an inlined copy_from_user_nofault for < 5.8.
The old probe_user_write for < 5.8 is NOT really _nofault as it does NOT have that
pagefault_disable + enable which is the hallmark of these _nofault functions.

devlog:
https://github.com/ximi-libra-test/android_kernel_xiaomi_libra/compare/16e5dce9e7e...16c1f5f521a
https://github.com/backslashxx/mojito_krenol/compare/28642e60d7b...728de0c571c

...

-https://github.com/backslashxx/KernelSU/issues/5
__kernel_compat: add ksu_copy_from_user_retry
__ksud: remove is_locked_copy_ok and change to this in ksu_handle_pre_ksud
-https://gitlab.com/pershoot/susfs4ksu/-/tree/gki-android14-6.1-dev